+++ src/constants/index.js
@@ -1,0 +1,16 @@
+import * as CONFIG from './config.js';
+import { PRODUCTS_ROUTE, PROMOTIONS_ROUTE } from './route.js';
+import * as ERROR_MESSAGES from './errorMessages.js';
+import { MESSAGES, ORDER_MESSAGES, RECEIPT_MESSAGES } from './messages.js';
+
+export { default as INITIAL_PRODUCTS } from './initialProducts.js';
+
+export {
+  CONFIG,
+  PRODUCTS_ROUTE,
+  PROMOTIONS_ROUTE,
+  ERROR_MESSAGES,
+  MESSAGES,
+  ORDER_MESSAGES,
+  RECEIPT_MESSAGES,
+};
