+++ src/constants/messages.js
@@ -5,7 +5,7 @@ export const MESSAGES = Object.freeze({
   askToContinue: '\n감사합니다. 구매하고 싶은 다른 상품이 있나요? (Y/N)\n',
 });
 
-export const orderMessages = Object.freeze({
+export const ORDER_MESSAGES = Object.freeze({
   askAddItem({ name }) {
     return `\n현재 ${name}은(는) 1개를 무료로 더 받을 수 있습니다. 추가하시겠습니까? (Y/N)\n`;
   },
