+++ src/io/index.js
@@ -1,0 +1,5 @@
+import FileSystem from './FileSystem.js';
+import Printer from './Printer.js';
+import Reader from './Reader.js';
+
+export { FileSystem, Printer, Reader };
