+++ src/models/index.js
@@ -1,0 +1,5 @@
+import Promotion from './Promotion.js';
+import Receipt from './Receipt.js';
+import Shelves from './Shelves.js';
+
+export { Promotion, Receipt, Shelves };
