+++ src/services/index.js
@@ -1,0 +1,5 @@
+import AskHandler from './AskHandler.js';
+import OrderHandler from './OrderHandler.js';
+import PurchaseService from './PurchaseService.js';
+
+export { AskHandler, OrderHandler, PurchaseService };
