+++ src/view/OutputView.js
@@ -1,7 +1,6 @@
-import { MESSAGES, RECEIPT_MESSAGES } from '../constants/messages.js';
-import Printer from '../io/Printer.js';
-import * as format from '../utils/formatters.js';
-import * as unit from '../utils/units.js';
+import { MESSAGES, RECEIPT_MESSAGES } from '../constants/index.js';
+import { Printer } from '../io/index.js';
+import { format, unit } from '../utils/index.js';
 
 class OutputView {
   static printIntroduce() {
